+++ prof-designs-guardian.php
@@ -3,7 +3,7 @@
      * Plugin Name: Prof Designs Guardian
      * Plugin URI: https://prof-designs.com/guardian
      * Description: A  plugin that provides automatic updates, error handling, and health checks for your website.
-     * Version: 0.1.1
+     * Version: 0.2.0
      *
      * Author: Prof Designs
      * Author URI: https://profdesigns.com
@@ -25,10 +25,13 @@
 
     require_once __DIR__ . '/includes/Helpers.php';
     require_once __DIR__ . '/includes/Mailer.php';
+
+    require_once __DIR__ . '/includes/Security.php';
     require_once __DIR__ . '/includes/AutoUpdates.php';
     require_once __DIR__ . '/includes/ErrorHandler.php';
     require_once __DIR__ . '/includes/HealthCheck.php';
 
+    ProfDesigns\Guardian\Security::init();
     ProfDesigns\Guardian\AutoUpdates::init();
     ProfDesigns\Guardian\ErrorHandler::init();
     ProfDesigns\Guardian\HealthCheck::init();